requirements.txt
# Deep Learning for Solving and Estimating Dynamic Models in Economics and Finance
# Python dependencies for self-study readers running notebooks locally.
# Course platform (Nuvolos Cloud) ships these pre-installed.
#
# Install: pip install -r requirements.txt
# Python 3.10+ required.

# Core scientific computing
numpy>=1.24
scipy>=1.10
pandas>=2.0

# Visualization
matplotlib>=3.7

# Classical machine learning
scikit-learn>=1.3

# Deep learning frameworks
tensorflow>=2.15
tensorflow-probability>=0.23
torch>=2.0

# JAX stack (Krusell-Smith JAX port, sequence-space DEQNs, selected GP routines)
jax>=0.4.20
jaxlib>=0.4.20
optax>=0.1.5
flax>=0.7.0

# Gaussian process regression (Lectures 23-25)
gpytorch>=1.11
botorch>=0.11

# Training monitoring
tensorboard>=2.15

# Utilities
requests>=2.28
pyyaml>=6.0
